@@ -1,15 +1,27 @@
 class TournamentsController < ApplicationController
   before_action :set_tournament, except: [:index, :new, :create]
-  before_action :check_if_admin, except: [:index, :show, :add_player, :remove_player, :location]
+  before_action :authenticate_admin!, except: [:index, :show, :add_player, :remove_player]
   before_action { @section = 'tournaments' }
 
   # GET /tournaments
   # GET /tournaments.json
   def index
-    @tournaments = Tournament.active_2019.upcoming.order(date: :asc).includes(:players).paginate(page: params[:page], per_page: Tournament::MAX_PAST_TOURNAMENTS_PER_PAGE)
-    @ongoing_tournaments = Tournament.active_2019.ongoing.order(date: :asc).includes(:players).paginate(page: params[:page], per_page: Tournament::MAX_PAST_TOURNAMENTS_PER_PAGE)
-    @past_tournaments = Tournament.active_2019.past.order(date: :desc).includes(:players).paginate(page: params[:page], per_page: Tournament::MAX_PAST_TOURNAMENTS_PER_PAGE)
-    if params[:filter].present?
+    @tournaments = Tournament.all_from(session['country_code']).active.upcoming.order(date: :asc).includes(:players).paginate(page: params[:page], per_page: Tournament::MAX_PAST_TOURNAMENTS_PER_PAGE)
+    @ongoing_tournaments = Tournament.all_from(session['country_code']).active.ongoing.order(date: :asc).includes(:players).paginate(page: params[:page], per_page: Tournament::MAX_PAST_TOURNAMENTS_PER_PAGE)
+    @past_tournaments = Tournament.all_from(session['country_code']).active.past.order(date: :desc).includes(:players).paginate(page: params[:page], per_page: Tournament::MAX_PAST_TOURNAMENTS_PER_PAGE)
+    if current_user.present? and current_user.super_admin?
+      @inactive_tournaments = Tournament.all_from(session['country_code']).where(active: false).order(date: :desc).paginate(page: params[:page], per_page: Tournament::MAX_PAST_TOURNAMENTS_PER_PAGE)
+    end
+    # handle search parameter
+    if params[:search].present?
+      @tournaments = @tournaments.search(params[:search])
+      @past_tournaments = @past_tournaments.search(params[:search])
+      if @tournaments.empty? and @past_tournaments.empty?
+        flash.now[:alert] = t('flash.alert.search_tournaments')
+      end
+      @inactive_tournaments = @inactive_tournaments.search(params[:search]) if @inactive_tournaments.present?
+    end
+    if params[:filter].present? and params[:filter] != 'all'
       if helpers.tournament_cities.include?(params[:filter].capitalize)
         city = params[:filter].capitalize
         @tournaments = @tournaments.where(city: city).or(
@@ -18,36 +30,65 @@ class TournamentsController < ApplicationController
         @past_tournaments = @past_tournaments.where(city: city).or(
           @past_tournaments.from_city(city)
         )
+      elsif helpers.federal_states_raw.include?(params[:filter].upcase)
+        federal_state = params[:filter].upcase
+        @tournaments = @tournaments.where(federal_state: federal_state)
+        @past_tournaments = @past_tournaments.where(federal_state: federal_state)
+      elsif helpers.regions_raw.include?(params[:filter])
+        region = params[:filter]
+        @tournaments = @tournaments.where(region: region)
+        @past_tournaments = @past_tournaments.where(region: region)
+      elsif params[:filter] == 's1_2019'
+        @tournaments = @tournaments.where('date >= ? AND date < ?', Time.local(2019,1,1), Time.local(2019,6,16)).where.not(subtype: 'weekly').where("name NOT ILIKE ?", "%Weekly%")
+        @past_tournaments = @past_tournaments.where('date >= ? AND date < ?', Time.local(2019,1,1), Time.local(2019,6,16)).where.not(subtype: 'weekly').where("name NOT ILIKE ?", "%Weekly%")
+      elsif params[:filter] == 's2_2019'
+        @tournaments = @tournaments.where('date >= ? AND date < ?', Time.local(2019,6,16), Time.local(2019,12,31,23,59,59)).where.not(subtype: 'weekly').where("name NOT ILIKE ?", "%Weekly%")
+        @past_tournaments = @past_tournaments.where('date >= ? AND date < ?', Time.local(2019,6,16), Time.local(2019,12,31,23,59,59)).where.not(subtype: 'weekly').where("name NOT ILIKE ?", "%Weekly%")
+      elsif params[:filter] == 's12_2020'
+        @tournaments = @tournaments.where('date >= ? AND date < ?', Time.local(2020,1,1), Time.local(2020,12,31,23,59,59)).where.not(subtype: 'weekly').where("name NOT ILIKE ?", "%Weekly%")
+        @past_tournaments = @past_tournaments.where('date >= ? AND date < ?', Time.local(2020,1,1), Time.local(2020,12,31,23,59,59)).where.not(subtype: 'weekly').where("name NOT ILIKE ?", "%Weekly%")
+      elsif params[:filter] == 's12_2021'
+        @tournaments = @tournaments.where('date >= ? AND date < ?', Time.local(2021,1,1), Time.local(2021,12,31,23,59,59)).where.not(subtype: 'weekly').where("name NOT ILIKE ?", "%Weekly%")
+        @past_tournaments = @past_tournaments.where('date >= ? AND date < ?', Time.local(2021,1,1), Time.local(2021,12,31,23,59,59)).where.not(subtype: 'weekly').where("name NOT ILIKE ?", "%Weekly%")
+      elsif params[:filter] == 's12_2022'
+        @tournaments = @tournaments.where('date >= ? AND date < ?', Time.local(2022,1,1), Time.local(2022,12,31,23,59,59)).where.not(subtype: 'weekly').where("name NOT ILIKE ?", "%Weekly%")
+        @past_tournaments = @past_tournaments.where('date >= ? AND date < ?', Time.local(2022,1,1), Time.local(2022,12,31,23,59,59)).where.not(subtype: 'weekly').where("name NOT ILIKE ?", "%Weekly%")
       else  # 'weekly', 'internal' or 'external'
         @tournaments = @tournaments.where(subtype: params[:filter])
         @past_tournaments = @past_tournaments.where(subtype: params[:filter])
       end
-    end
-    if current_user.present? and current_user.super_admin?
-      @inactive_tournaments = Tournament.where(active: false).order(date: :desc).paginate(page: params[:page], per_page: Tournament::MAX_PAST_TOURNAMENTS_PER_PAGE)
     end
   end
 
   # GET /tournaments/1
   # GET /tournaments/1.json
   def show
-    players_per_game_station = (@tournament.total_seats.to_f/@tournament.total_needed_game_stations).ceil() if @tournament.total_needed_game_stations.to_i > 0
-    @min_needed_game_stations = (@tournament.min_needed_registrations.to_f/players_per_game_station).ceil() if players_per_game_station.to_i > 0
-    if @tournament.players.count < @tournament.min_needed_registrations.to_i
-      @currently_needed_game_stations = @min_needed_game_stations - @tournament.game_stations_count if @min_needed_game_stations.present?
-    else
-      @currently_needed_game_stations = (@tournament.players.count.to_f/players_per_game_station).ceil() - @tournament.game_stations_count if players_per_game_station.to_i > 0
-    end
-    @registration = @tournament.registrations.where(player_id: current_user.player.id).first if current_user.present?
+    @currently_needed_game_stations = @tournament.total_needed_game_stations - @tournament.game_stations_count if @tournament.total_needed_game_stations.present?
+    @registration = @tournament.registrations.find_by(player_id: current_user.player.id) if current_user.present?
+    @seeded_participants_array = @tournament.registrations.order(:position).map { |r| r.player.gamer_tag }
+    @seeded_participants_array = reorder_for_pools(@seeded_participants_array, @tournament.number_of_pools) if @tournament.has_pools?
   end
 
   # GET /tournaments/new
   def new
-    @tournament = Tournament.new
+    if params[:id].present?
+      @tournament = Tournament.find(params[:id]).dup
+      @tournament.active = true
+      @tournament.started = false
+      @tournament.finished = false
+      @tournament.challonge_tournament_id = nil
+      @tournament.ranking_string = nil
+      @tournament.setup = false
+      @tournament.waiting_list = []
+      @tournament.external_registration_link = nil
+    else
+      @tournament = Tournament.new(params[:tournament].present? ? tournament_params : nil)
+    end
   end
 
   # GET /tournaments/1/edit
   def edit
+    @tournament.assign_attributes(tournament_params) if params[:tournament].present?
   end
 
   # POST /tournaments
@@ -55,21 +96,22 @@ class TournamentsController < ApplicationController
   def create
     @tournament = Tournament.new(tournament_params)
     @tournament.ranking_string = ''
+    @tournament.country_code = session['country_code']
     # handle the different subtypes
     if @tournament.subtype.nil? or @tournament.subtype == 'internal'
       respond_to do |format|
         if check_registration_deadline_is_less_than_date(tournament_params) && @tournament.save
           if params[:send_mails]
-            Player.all.each do |p|
-              if p.user.wants_major_email
+            Player.all_from(session['country_code']).each do |p|
+              if p.user.allows_emails
                 TournamentMailer.with(tournament: @tournament, user: p.user).new_tournament_email.deliver_later
               end
             end
           end
-          format.html { redirect_to @tournament, notice: 'Internal tournament was successfully created.' }
+          format.html { redirect_to @tournament, notice: t('flash.notice.create_internal_tournament') }
           format.json { render :show, status: :created, location: @tournament }
         else
-          format.html { render :new }
+          format.html { redirect_to new_tournament_path(@tournament, tournament: tournament_params, errors: {count: @tournament.errors&.count, full_messages: @tournament.errors&.full_messages}, anchor: @tournament.subtype), alert: t('flash.alert.create_internal_tournament') }
           format.json { render json: @tournament.errors, status: :unprocessable_entity }
         end
       end
@@ -77,25 +119,25 @@ class TournamentsController < ApplicationController
       respond_to do |format|
         if @tournament.save
           if params[:send_mails]
-            Player.all.each do |p|
-              if p.user.wants_major_email
+            Player.all_from(session['country_code']).each do |p|
+              if p.user.allows_emails
                 TournamentMailer.with(tournament: @tournament, user: p.user).new_external_tournament_email.deliver_later
               end
             end
           end
-          format.html { redirect_to tournaments_path, notice: 'External tournament was successfully created.' }
+          format.html { redirect_to tournaments_path, notice: t('flash.notice.create_external_tournament') }
           format.json { render :show, status: :created, location: @tournament }
         else
-          format.html { render :new }
+          format.html { redirect_to new_tournament_path(@tournament, tournament: tournament_params, errors: {count: @tournament.errors&.count, full_messages: @tournament.errors&.full_messages}, anchor: @tournament.subtype), alert: t('flash.alert.create_external_tournament') }
           format.json { render json: @tournament.errors, status: :unprocessable_entity }
         end
       end
-    elsif @tournament.subtype == 'weekly'
+    elsif @tournament.weekly?
       @tournament.name = generate_weekly_name(@tournament.city, @tournament.date)
       respond_to do |format|
         if check_registration_deadline_is_less_than_date(tournament_params) && @tournament.save
           if params[:send_mails]
-            Player.all.each do |p|
+            Player.all_from(session['country_code']).each do |p|
               if p.user.wants_weekly_email
                 TournamentMailer.with(tournament: @tournament, user: p.user).new_weekly_tournament_email.deliver_later
               end
@@ -111,15 +153,15 @@ class TournamentsController < ApplicationController
             if next_weekly.save
               last_weekly = next_weekly
             else
-              format.html { render :new }
+              format.html { redirect_to new_tournament_path(@tournament, tournament: tournament_params, errors: {count: @tournament.errors&.count, full_messages: @tournament.errors&.full_messages}, anchor: @tournament.subtype), alert: t('flash.alert.create_weekly_tournament') }
               format.json { render json: next_weekly.errors, status: :unprocessable_entity }
               return
             end
           end
-          format.html { redirect_to @tournament, notice: 'One or more weeklies were successfully created.' }
+          format.html { redirect_to @tournament, notice: t('flash.notice.create_weekly_tournament') }
           format.json { render :show, status: :created, location: @tournament }
         else
-          format.html { render :new }
+          format.html { redirect_to new_tournament_path(@tournament, tournament: tournament_params, errors: {count: @tournament.errors&.count, full_messages: @tournament.errors&.full_messages}, anchor: @tournament.subtype), alert: t('flash.alert.create_weekly_tournament') }
           format.json { render json: @tournament.errors, status: :unprocessable_entity }
         end
       end
@@ -135,28 +177,28 @@ class TournamentsController < ApplicationController
   # PATCH/PUT /tournaments/1.json
   def update
     # update tournament
-    if @tournament.subtype.nil? or @tournament.subtype == 'internal' or  @tournament.subtype == 'weekly'
+    if @tournament.subtype.nil? or @tournament.subtype == 'internal' or @tournament.weekly?
       respond_to do |format|
         oldName = @tournament.name
         oldDate = @tournament.date
         oldcity = @tournament.city
         if check_registration_deadline_is_less_than_date(tournament_params) && @tournament.update(tournament_params)
           # also update weekly name if city was edited
-          if @tournament.subtype == 'weekly' && @tournament.city != oldcity
+          if @tournament.weekly? && @tournament.city != oldcity
             @tournament.name = generate_weekly_name(@tournament.city, @tournament.date)
             unless @tournament.save
-              format.html { render :edit }
+              format.html { redirect_to edit_tournament_path(@tournament, tournament: tournament_params, errors: {count: @tournament.errors&.count, full_messages: @tournament.errors&.full_messages}, anchor: @tournament.subtype), alert: t('flash.alert.update_weekly_tournament') }
               format.json { render json: t.errors, status: :unprocessable_entity }
               return
             end
           end
 
           # check the 'all' parameter and update all upcoming tournaments of this type if it's true
-          if @tournament.subtype == 'weekly' and params[:commit].include?('all')
+          if @tournament.weekly? and params[:commit].include?('all')
             last_weekly = @tournament
             edited_tournament_params = tournament_params
             old_name_without_kw = oldName[0.. -10].strip  # 'SSBU Weekly xxx KWyy 20zz' -> 'SSBU Weekly xxx'
-            Tournament.where('date >= ?', oldDate + 7.days).where("name ILIKE ?", "%#{ActiveRecord::Base.sanitize_sql_like(old_name_without_kw)}%").order(:date).each do |t|
+            Tournament.all_from(session['country_code']).where('date >= ?', oldDate + 7.days).where("name ILIKE ?", "%#{ActiveRecord::Base.sanitize_sql_like(old_name_without_kw)}%").order(:date).each do |t|
               edited_tournament_params[:name] = generate_weekly_name(last_weekly.city, last_weekly.date + 7.days)
               if t.update(edited_tournament_params)
                 t.date = last_weekly.date + 7.days
@@ -164,33 +206,33 @@ class TournamentsController < ApplicationController
                 if t.save
                   last_weekly = t
                 else
-                  format.html { render :edit }
+                  format.html { redirect_to edit_tournament_path(@tournament, tournament: tournament_params, errors: {count: @tournament.errors&.count, full_messages: @tournament.errors&.full_messages}, anchor: @tournament.subtype), alert: t('flash.alert.update_weekly_tournament') }
                   format.json { render json: t.errors, status: :unprocessable_entity }
                   return
                 end
               else
-                format.html { render :edit }
+                format.html { redirect_to edit_tournament_path(@tournament, tournament: tournament_params, errors: {count: @tournament.errors&.count, full_messages: @tournament.errors&.full_messages}, anchor: @tournament.subtype), alert: t('flash.alert.update_weekly_tournament') }
                 format.json { render json: t.errors, status: :unprocessable_entity }
                 return
               end
             end
-            format.html { redirect_to @tournament, notice: 'One or more weeklies were successfully updated.' }
+            format.html { redirect_to @tournament, notice: t('flash.notice.update_weekly_tournament') }
           else
-            format.html { redirect_to @tournament, notice: 'Tournament was successfully updated.' }
+            format.html { redirect_to @tournament, notice: t('flash.notice.update_internal_tournament') }
           end
           format.json { render :show, status: :ok, location: @tournament }
         else
-          format.html { render :edit }
+          format.html { redirect_to edit_tournament_path(@tournament, tournament: tournament_params, errors: {count: @tournament.errors&.count, full_messages: @tournament.errors&.full_messages}, anchor: @tournament.subtype), alert: t('flash.alert.update_internal_tournament') }
           format.json { render json: @tournament.errors, status: :unprocessable_entity }
         end
       end
     elsif @tournament.subtype == 'external'
       respond_to do |format|
         if check_registration_deadline_is_less_than_date(tournament_params) && @tournament.update(tournament_params)
-          format.html { redirect_to tournaments_path, notice: 'External tournament was successfully updated.' }
+          format.html { redirect_to tournaments_path, notice: t('flash.notice.update_external_tournament') }
           format.json { render :show, status: :ok, location: @tournament }
         else
-          format.html { render :edit }
+          format.html { redirect_to edit_tournament_path(@tournament, tournament: tournament_params, errors: {count: @tournament.errors&.count, full_messages: @tournament.errors&.full_messages}, anchor: @tournament.subtype), alert: t('flash.alert.update_external_tournament') }
           format.json { render json: @tournament.errors, status: :unprocessable_entity }
         end
       end
@@ -205,89 +247,120 @@ class TournamentsController < ApplicationController
   # DELETE /tournaments/1
   # DELETE /tournaments/1.json
   def destroy
-    if @tournament.subtype == 'weekly' and params[:all]
-      # deactivate all upcoming weeklies of this type
-      name_without_kw = @tournament.name[0.. -10].strip  # 'SSBU Weekly xxx KWyy 20zz' -> 'SSBU Weekly xxx'
-      Tournament.where('date >= ?', @tournament.date).where(location: @tournament.location).where(host_username: @tournament.host_username).each do |tt|
-        if tt.name[0.. -10].strip == name_without_kw
-          tt.update(active: false)
+    if @tournament.active == false
+      if @tournament.destroy
+        respond_to do |format|
+          format.html { redirect_to tournaments_url, notice: t('flash.notice.delete_tournament') }
+          format.json { head :no_content }
+        end
+      else
+        redirect_to tournaments_url, alert: t('flash.alert.delete_tournament')
+      end
+    else
+      if @tournament.weekly? and params[:all]
+        # deactivate all upcoming weeklies of this type
+        name_without_kw = @tournament.name[0.. -10].strip  # 'SSBU Weekly xxx KWyy 20zz' -> 'SSBU Weekly xxx'
+        Tournament.all_from(session['country_code']).where('date >= ?', @tournament.date).where(location: @tournament.location).where(host_username: @tournament.host_username).each do |tt|
+          if tt.name[0.. -10].strip == name_without_kw
+            tt.update(active: false)
+          end
         end
       end
-    end
-    @tournament.update(active: false)
-    respond_to do |format|
-      format.html { redirect_to tournaments_url, notice: 'Tournament was successfully deleted.' }
-      format.json { head :no_content }
+      if @tournament.update(active: false)
+        respond_to do |format|
+          format.html { redirect_to tournaments_url, notice: t('flash.notice.deactivate_tournament') }
+          format.json { head :no_content }
+        end
+      else
+        redirect_to tournaments_url, alert: t('flash.alert.deactivate_tournament')
+      end
     end
   end
 
   # POST /tournaments/add_player/1
   def add_player
-    if params[:gamer_tag].present?
-      player_to_add = Player.find_by(gamer_tag: params[:gamer_tag])
-    else
-      player_to_add = current_user.player
+    if current_user.present?
+      if params[:gamer_tag].present? && current_user.admin?
+        player_to_add = Player.find_by(gamer_tag: params[:gamer_tag])
+        player_to_add = AlternativeGamerTag.find_by(gamer_tag: params[:gamer_tag]).try(:player) if player_to_add.nil?
+      else
+        player_to_add = current_user.player
+      end
     end
 
     if player_to_add.nil?
-      redirect_to @tournament, alert: "Player couldn't be added to the tournament -> Player not found."
+      redirect_to @tournament, alert: "#{t('flash.alert.add_player_failed')} -> #{t('flash.alert.player_not_found')}"
       return;
     end
 
     if @tournament.registration_deadline and Time.now > @tournament.registration_deadline and !params[:gamer_tag].present?
-      redirect_to @tournament, alert: "Player couldn't be added to the tournament -> Online registration deadline exceeded."
+      redirect_to @tournament, alert: "#{t('flash.alert.add_player_failed')} -> #{t('flash.alert.deadline_exceeded')}"
       return;
     end
 
     if @tournament.players.include?(player_to_add)
-      redirect_to @tournament, alert: "Player couldn't be added to the tournament -> Player was already added."
+      redirect_to @tournament, alert: "#{t('flash.alert.add_player_failed')} -> #{t('flash.alert.player_already_added')}"
       return;
     end
 
     if @tournament.total_seats.present? and @tournament.players.count < @tournament.total_seats
       # tournament is not full yet -> add the player to the tournament
       @tournament.players << player_to_add
+      # re-seed players if not added manually
+      tournament_registrations = @tournament.registrations
+      if params[:gamer_tag].present?
+        tournament_registrations.last.set_list_position(@tournament.registrations.count)
+      else
+        seeded_participants_id_array = helpers.seed_players(@tournament.players).map { |p| p.id }
+        seeded_participants_id_array.each_with_index do |id, i|
+          tournament_registrations.find_by(player_id: id).set_list_position(i+1)
+        end
+      end
       # remove the player from the waiting list
       if @tournament.waiting_list.include?(player_to_add.gamer_tag)
         @tournament.waiting_list.delete(player_to_add.gamer_tag)
         if @tournament.save
-          redirect_to @tournament, notice: 'Player was added to the tournament and removed from the waiting list.'
+          redirect_to @tournament, notice: "#{t('flash.notice.add_player')} #{t('flash.notice.removed_from_wating_list')}"
         else
-          redirect_to @tournament, alert: "Player was added to the tournament but couldn't be removed from the waiting list."
+          redirect_to @tournament, alert: "#{t('flash.notice.add_player')} #{t('flash.alert.not_removed_from_wating_list')}"
         end
       else
-        redirect_to @tournament, notice: 'Player was added to the tournament.'
+        redirect_to @tournament, notice: t('flash.notice.add_player')
       end
     else
       # tournament is full
-      if params[:waiting_list] == 'true' or (params[:gamer_tag].present? and !@tournament.players.include?(player_to_add))
+      if (params[:waiting_list] == 'true' and !@tournament.waiting_list.include?(player_to_add.gamer_tag)) or (params[:gamer_tag].present? and !@tournament.players.include?(player_to_add) and !@tournament.waiting_list.include?(player_to_add.gamer_tag))
         @tournament.waiting_list.push(player_to_add.gamer_tag)
         if @tournament.save
-          redirect_to @tournament, notice: "Player was added to the waiting list."
+          redirect_to @tournament, notice: t('flash.notice.add_player_to_waiting_list')
         else
-          redirect_to @tournament, alert: "Player couldn't be added to the waiting list."
+          redirect_to @tournament, alert: t('flash.alert.add_player_to_waiting_list')
         end
       else
-        redirect_to @tournament, alert: "Player couldn't be added to the tournament -> The tournament is full."
+        redirect_to @tournament, alert: "#{t('flash.alert.add_player_failed')} -> #{t('flash.alert.tournament_full')}"
       end
     end
   end
 
   # POST /tournaments/remove_player/1
   def remove_player
-    if params[:gamer_tag].present?
-      player_to_remove = Player.find_by(gamer_tag: params[:gamer_tag])
-    else
-      player_to_remove = current_user.player
+    if current_user.present?
+      if params[:gamer_tag].present? && current_user.admin?
+        player_to_remove = Player.find_by(gamer_tag: params[:gamer_tag])
+        player_to_remove = AlternativeGamerTag.find_by(gamer_tag: params[:gamer_tag]).try(:player) if player_to_remove.nil?
+        player_to_remove.update(warnings: player_to_remove.warnings.to_i + 1) if params[:warn].present?
+      else
+        player_to_remove = current_user.player
+      end
     end
 
     if player_to_remove.nil?
-      redirect_to @tournament, alert: "Player couldn't be removed from the tournament -> Player not found."
+      redirect_to @tournament, alert: "#{t('flash.alert.add_player_failed')} -> #{t('flash.alert.player_not_found')}"
       return;
     end
 
     if @tournament.registration_deadline and Time.now > @tournament.registration_deadline and !params[:gamer_tag].present?
-      redirect_to @tournament, alert: "Player couldn't be removed from the tournament -> Online registration deadline exceeded."
+      redirect_to @tournament, alert: "#{t('flash.alert.add_player_failed')} -> #{t('flash.alert.deadline_exceeded')}"
       return;
     end
 
@@ -295,12 +368,12 @@ class TournamentsController < ApplicationController
       if @tournament.waiting_list.include?(player_to_remove.gamer_tag)
         @tournament.waiting_list.delete(player_to_remove.gamer_tag)
         if @tournament.save
-          redirect_to @tournament, notice: 'Player was removed from the waiting list.'
+          redirect_to @tournament, notice: t('flash.notice.remove_player_from_waiting_list')
         else
-          redirect_to @tournament, alert: "Player couldn't be removed from the waiting list."
+          redirect_to @tournament, alert: t('flash.alert.remove_player_from_waiting_list')
         end
       else
-        redirect_to @tournament, alert: "Player couldn't be removed from the tournament -> Player is not in the tournament."
+        redirect_to @tournament, alert: "#{t('flash.alert.remove_player_failed')} -> #{t('flash.alert.player_not_in_tournament')}"
       end
       return;
     end
@@ -312,96 +385,87 @@ class TournamentsController < ApplicationController
     if @tournament.waiting_list.length > 0
       gamer_tag_to_add = @tournament.waiting_list.shift # shift is the same as pop_first
       player_to_add = Player.find_by(gamer_tag: gamer_tag_to_add)
+      player_to_add = AlternativeGamerTag.find_by(gamer_tag: gamer_tag_to_add).try(:player) if player_to_add.nil?
       @tournament.players << player_to_add
       TournamentMailer.with(tournament: @tournament, user: player_to_add.user).waiting_player_upgraded_email.deliver_later
       if @tournament.save
-        redirect_to @tournament, notice: 'Player was removed from the tournament and the first player in the waiting list took the seat.'
+        redirect_to @tournament, notice: "#{t('flash.notice.remove_player')} #{t('flash.notice.first_waiting_player')}"
       else
-        redirect_to @tournament, alert: "Player was removed from the tournament but the first player in the waiting list couldn't take the seat."
+        redirect_to @tournament, alert: "#{t('flash.notice.remove_player')} #{t('flash.alert.first_waiting_player')}"
       end
     else
-      redirect_to @tournament, notice: 'Player was removed from the tournament.'
+      redirect_to @tournament, notice: t('flash.notice.remove_player')
     end
   end
 
   # POST /tournaments/setup/1'
   def setup
     if @tournament.setup or @tournament.started or @tournament.finished
-      redirect_to @tournament, alert: 'Tournament is already set up, started or finished.'
+      redirect_to @tournament, alert: t('flash.alert.tournament_status_error')
     else
-      min_needed_game_stations_count = helpers.min_needed_game_stations_per_tournament(@tournament.players.count)
-      current_game_stations_count = @tournament.game_stations_count
-      if current_game_stations_count < min_needed_game_stations_count
-        delta_game_stations = min_needed_game_stations_count - current_game_stations_count
-        redirect_to @tournament, alert: "At least #{delta_game_stations} more game #{delta_game_stations > 1 ? 'stations are' : 'station is'} needed to setup the tournament."
-      else
-        if set_challonge_username_and_api_key()
+      if set_challonge_username_and_api_key()
 
-          # setup a challonge tournament
-          ct = Challonge::Tournament.new
-          ct.name = @tournament.name #'SSBU Bern KW1' or 'PK Bern #1'
-          ct.url = helpers.valid_challonge_url(@tournament.name) #'ssbu_bern_kw1' or 'pk_bern_1'
-          ct.tournament_type = 'double elimination'
-          ct.group_stages_enabled = @tournament.has_pools?
-          ct.game_name = 'Super Smash Bros. Ultimate'
-          ct.description = @tournament.description
-          if ct.save == false
+        # setup a challonge tournament
+        ct = Challonge::Tournament.new
+        ct.name = @tournament.name #'SSBU Bern KW1' or 'PK Bern #1'
+        ct.url = helpers.valid_challonge_url(@tournament.name) #'ssbu_bern_kw1' or 'pk_bern_1'
+        ct.tournament_type = 'double elimination'
+        ct.group_stages_enabled = @tournament.has_pools?
+        ct.game_name = 'Super Smash Bros. Ultimate'
+        ct.description = helpers.valid_challonge_description(@tournament.description)
+        ActiveRecord::Base.transaction do
+          if !ct.save
             raise ct.errors.full_messages.inspect
           end
+        rescue => error
+          redirect_to @tournament, alert: error
+          return
+        end
 
-          # sort the participants by the best player
-          seeded_participants_array = @tournament.players.sort_by do |p|
-            p.seed_points
-          end.reverse.map { |p| p.gamer_tag }
+        # get seeded players
+        seeded_participants_array = []
+        all_registrations = @tournament.registrations
+        if all_registrations.where(position: nil).any?
+          # use seed_points if a position is nil
+          seeded_participants_array = helpers.seed_players(@tournament.players).map { |p| p.gamer_tag }
+        else
+          seeded_participants_array = all_registrations.order(:position).map { |r| r.player.gamer_tag }
+        end
 
-          # change the order if there are pools
-          if @tournament.has_pools?
-            num_of_pools = @tournament.number_of_pools
-            players_per_pool = (seeded_participants_array.size.to_f/num_of_pools).round
-            pools_hash = Hash.new
-            num_of_pools.times do |n|
-              pools_hash[n] = Array.new
-            end
-            i = 0
-            (players_per_pool.to_f/2).round.times do
-              num_of_pools.times do |n|
-                pools_hash[n%num_of_pools] << seeded_participants_array[i]
-                i += 1
-              end
-              num_of_pools.times do |n|
-                pools_hash[(num_of_pools-1-n)%num_of_pools] << seeded_participants_array[i]
-                i += 1
-              end
-            end
-            seeded_participants_array = []
-            num_of_pools.times do |n|
-              seeded_participants_array += pools_hash[n]
-              if seeded_participants_array.include?(nil) && n != (num_of_pools-1)
-                seeded_participants_array.pop # remove the nil at the end
-                seeded_participants_array << pools_hash[(num_of_pools-1)].pop # add the last player
-              end
-            end
-          end
+        # change the order if there are pools
+        seeded_participants_array = reorder_for_pools(seeded_participants_array, @tournament.number_of_pools) if @tournament.has_pools?
 
-          # add the participants to the challonge tournament
-          seeded_participants_array.each do |p|
-            Challonge::Participant.create(:name => p, :tournament => ct)
-          end
-
-          @tournament.setup = true
-          @tournament.challonge_tournament_id = ct.id
-          if @tournament.save
-            if @tournament.has_pools?
-              redirect_to @tournament, notice: "Tournament was successfully set up. This is a two-stage tournament -> Check out the preferences on challonge.com to configure your pools."
-            else
-              redirect_to @tournament, notice: "Tournament was successfully set up. Check it out on challonge.com and click 'Start tournament' if you're ready."
-            end
+        # add "bye"-players if there are nils
+        bye_ctr = 0
+        seeded_participants_array = seeded_participants_array.map do |p|
+          if p.nil?
+            bye_gamer_tag = "bye#{bye_ctr}"
+            bye_ctr += 1
+            @tournament.players << Player.find_by(gamer_tag: bye_gamer_tag)
+            bye_gamer_tag
           else
-            redirect_to @tournament, alert: "Tournament couldn't be set up."
+            p
+          end
+        end
+
+        # add the participants to the challonge tournament
+        seeded_participants_array.each do |p|
+          Challonge::Participant.create(:name => p, :tournament => ct)
+        end
+
+        @tournament.setup = true
+        @tournament.challonge_tournament_id = ct.id
+        if @tournament.save
+          if @tournament.has_pools?
+            redirect_to @tournament, notice: "#{t('flash.notice.tournament_set_up')}. #{t('flash.notice.two_stage_tournament')}"
+          else
+            redirect_to @tournament, notice: "#{t('flash.notice.tournament_set_up')}. #{t('flash.notice.check_out_challonge')}"
           end
         else
-          redirect_to @tournament, alert: "Tournament cannot be set up. Challonge data are missing. Add them #{view_context.link_to('here', edit_user_registration_path, target: '_blank')}".html_safe
+          redirect_to @tournament, alert: t('flash.alert.tournament_set_up')
         end
+      else
+        redirect_to @tournament, alert: "#{t('flash.alert.tournament_set_up')}. #{t('flash.alert.challonge_data_missing', link: view_context.link_to(t('flash.alert.here'), edit_user_registration_path, target: '_blank')).html_safe}"
       end
     end
   end
@@ -409,9 +473,9 @@ class TournamentsController < ApplicationController
   # POST /tournaments/start/1
   def start
     if !@tournament.setup
-      redirect_to @tournament, alert: "Tournament wasn't set up yet."
+      redirect_to @tournament, alert: t('flash.alert.not_set_up_yet')
     elsif @tournament.started or @tournament.finished
-      redirect_to @tournament, alert: 'Tournament is already started or finished.'
+      redirect_to @tournament, alert: t('flash.alert.already_started')
     else
       if set_challonge_username_and_api_key()
 
@@ -421,12 +485,12 @@ class TournamentsController < ApplicationController
         ct.start!
         @tournament.started = true
         if @tournament.save
-          redirect_to @tournament, notice: 'Tournament was successfully started.'
+          redirect_to @tournament, notice: t('flash.notice.tournament_started')
         else
-          redirect_to @tournament, alert: "Tournament couldn't be started."
+          redirect_to @tournament, alert: t('flash.alert.tournament_started')
         end
       else
-        redirect_to @tournament, alert: "Tournament cannot be started. Challonge data are missing. Add them #{view_context.link_to('here', edit_user_registration_path, target: '_blank')}".html_safe
+        redirect_to @tournament, alert: "#{t('flash.alert.tournament_started')}. #{t('flash.alert.challonge_data_missing', link: view_context.link_to(t('flash.alert.here'), edit_user_registration_path, target: '_blank')).html_safe}"
       end
     end
   end
@@ -434,9 +498,9 @@ class TournamentsController < ApplicationController
   # POST /tournaments/finish/1
   def finish
     if !@tournament.setup or !@tournament.started
-      redirect_to @tournament, alert: "Tournament wasn't set up or started yet."
+      redirect_to @tournament, alert: t('flash.alert.not_set_up_or_started_yet')
     elsif @tournament.finished
-      redirect_to @tournament, alert: 'Tournament is already finished.'
+      redirect_to @tournament, alert: t('flash.alert.tournament_finished')
     else
       if set_challonge_username_and_api_key()
 
@@ -449,85 +513,92 @@ class TournamentsController < ApplicationController
           ctms = ct.matches
 
           # updated the participated players and create the matches and results if no exception arises
-          ActiveRecord::Base.transaction do
-            # create matches
-            ctms.each do |ctm|
-              match = Match.new
-              match.challonge_match_id = ctm.id
-              match.tournament_id = @tournament.id
+          begin
+            ActiveRecord::Base.transaction do
+              # create matches
+              ctms.each do |ctm|
+                match = Match.new
+                match.challonge_match_id = ctm.id
+                match.tournament_id = @tournament.id
+                ctps.each do |ctp|
+                  gamer_tag = ctp.display_name.gsub("(invitation pending)", "").strip
+                  next if gamer_tag == 'bye0' || gamer_tag == 'bye1' || gamer_tag == 'bye2' || gamer_tag == 'bye3' || gamer_tag == 'bye4'
+                  if ctp.id == ctm.player1_id
+                    player = Player.find_by(gamer_tag: gamer_tag)
+                    player = AlternativeGamerTag.find_by(gamer_tag: gamer_tag).try(:player) if player.nil?
+                    raise ("#{ctp.display_name} not found in this tournament!").inspect if player.nil?
+                    match.player1_id = player.id
+                  elsif ctp.id == ctm.player2_id
+                    player = Player.find_by(gamer_tag: gamer_tag)
+                    player = AlternativeGamerTag.find_by(gamer_tag: gamer_tag).try(:player) if player.nil?
+                    raise ("#{ctp.display_name} not found in this tournament!").inspect if player.nil?
+                    match.player2_id = player.id
+                  end
+                end
+                scores = ctm.scores_csv.split('-')
+                match.player1_score = scores[0]
+                match.player2_score = scores[1]
+                match.save! # raise an exception when match.save failed
+              end
+
+              # create results and update players
               ctps.each do |ctp|
                 gamer_tag = ctp.display_name.gsub("(invitation pending)", "").strip
-                if ctp.id == ctm.player1_id
-                  player = Player.find_by(gamer_tag: gamer_tag)
-                  raise ("#{ctp.display_name} not found in this tournament!").inspect if player.nil?
-                  match.player1_id = player.id
-                elsif ctp.id == ctm.player2_id
-                  player = Player.find_by(gamer_tag: gamer_tag)
-                  raise ("#{ctp.display_name} not found in this tournament!").inspect if player.nil?
-                  match.player2_id = player.id
+                next if gamer_tag == 'bye0' || gamer_tag == 'bye1' || gamer_tag == 'bye2' || gamer_tag == 'bye3' || gamer_tag == 'bye4'
+                player = Player.find_by(gamer_tag: gamer_tag)
+                player = AlternativeGamerTag.find_by(gamer_tag: gamer_tag).try(:player) if player.nil?
+                raise ("#{ctp.display_name} not found in this tournament!").inspect if player.nil?
+                result = Result.new
+                result.player = player
+                result.tournament = @tournament
+                result.city = @tournament.city
+                result.rank = ctp.final_rank
+                result.points = helpers.points_repartition_table(ctp.final_rank)
+                player.points += result.points
+                player.participations += 1
+                if ctp.final_rank.present? and (player.best_rank == 0 or ctp.final_rank < player.best_rank) then player.best_rank = ctp.final_rank end
+                result.wins = 0
+                result.losses = 0
+                ctms.each do |ctm|
+                  scores = ctm.scores_csv.split('-')
+                  if ctm.player1_id == ctp.id
+                    result.wins += scores[0].to_i
+                    result.losses += scores[1].to_i
+                  elsif ctm.player2_id == ctp.id
+                    result.wins += scores[1].to_i
+                    result.losses += scores[0].to_i
+                  end
                 end
-              end
-              scores = ctm.scores_csv.split('-')
-              match.player1_score = scores[0]
-              match.player2_score = scores[1]
-              match.save! # raise an exception when match.save failed
-            end
-
-            # create results and update players
-            ctps.each do |ctp|
-              gamer_tag = ctp.display_name.gsub("(invitation pending)", "").strip
-              player = Player.find_by(gamer_tag: gamer_tag)
-              raise ("#{ctp.display_name} not found in this tournament!").inspect if player.nil?
-              result = Result.new
-              result.player = player
-              result.tournament = @tournament
-              result.city = @tournament.city
-              result.rank = ctp.final_rank
-              result.points = helpers.points_repartition_table(ctp.final_rank)
-              player.points += result.points
-              player.participations += 1
-              if ctp.final_rank.present? and (player.best_rank == 0 or ctp.final_rank < player.best_rank) then player.best_rank = ctp.final_rank end
-              result.wins = 0
-              result.losses = 0
-              ctms.each do |ctm|
-                scores = ctm.scores_csv.split('-')
-                if ctm.player1_id == ctp.id
-                  result.wins += scores[0].to_i
-                  result.losses += scores[1].to_i
-                elsif ctm.player2_id == ctp.id
-                  result.wins += scores[1].to_i
-                  result.losses += scores[0].to_i
+                player.wins += result.wins
+                player.losses += result.losses
+                if @tournament.subtype == 'internal'
+                  result.major_name = @tournament.name
                 end
-              end
-              player.wins += result.wins
-              player.losses += result.losses
-              if @tournament.subtype == 'internal'
-                result.major_name = @tournament.name
-              end
-              result.save! # raise an exception when result.save failed
-              player.save! # raise an exception when player.save failed
+                result.save! # raise an exception when result.save failed
+                player.save! # raise an exception when player.save failed
 
-              player.update_tournament_experience
+                player.update_tournament_experience
 
-              # updated raking_string on the tournament
-              ranking_string = "#{ctp.final_rank},#{gamer_tag};"
-              @tournament.ranking_string += ranking_string
+                # updated raking_string on the tournament
+                ranking_string = "#{ctp.final_rank},#{gamer_tag};"
+                @tournament.ranking_string += ranking_string
+              end
             end
+          rescue => error
+            redirect_to @tournament, alert: "#{t('flash.alert.tournament_cannot_finish')}.<br/><br/>Error:<br/>#{error}"
+            return
           end
-          # rescue ActiveRecord::RecordInvalid
-          #   puts "Something went wrong while updating the players!"
-          # end
 
           @tournament.finished = true
           @tournament.save
-          redirect_to @tournament, notice: 'Tournament was successfully finished and the participated players were updated.'
+          redirect_to @tournament, notice: t('flash.notice.tournament_finished')
         else
           if ct.started_at.nil? then @tournament.update(started: false) end # this happens when a ct was reset
           link = "https://challonge.com/#{ct.url}"
-          redirect_to @tournament, alert: "Tournament was not fineshed yet. You have to finish it first on: #{view_context.link_to link, link, target: '_blank'}".html_safe
+          redirect_to @tournament, alert: t('flash.alert.tournament_not_finished', link: view_context.link_to(link, link, target: '_blank')).html_safe
         end
       else
-        redirect_to @tournament, alert: "Tournament cannot be finished. Challonge data are missing. Add them #{view_context.link_to('here', edit_user_registration_path, target: '_blank')}".html_safe
+        redirect_to @tournament, alert: "#{t('flash.alert.tournament_cannot_finish')}. #{t('flash.alert.challonge_data_missing', link: view_context.link_to(t('flash.alert.here'), edit_user_registration_path, target: '_blank')).html_safe}"
       end
     end
   end
@@ -539,13 +610,26 @@ class TournamentsController < ApplicationController
       TournamentMailer.with(tournament: @tournament, user: p.user).tournament_cancelled_email.deliver_later
     end
     @tournament.update(tournament_params)
-    redirect_to @tournament, notice: 'Tournament was successfully cancelled.'
+    redirect_to @tournament, notice: t('flash.notice.tournament_cancelled')
   end
 
-  # GET /tournaments/location/1
-  # GET /tournaments/location/1.json
-  def location
+  # PATCH /tournaments/sort_players/1
+  def sort_players
+    params[:player].each_with_index do |id, i|
+      @tournament.registrations.find_by(player_id: id).update(position: i+1)
+    end
+    head :ok
+  end
 
+  # PATCH /tournaments/seed_players/1
+  def seed_players
+    tournament_registrations = @tournament.registrations
+    seeded_participants_id_array = helpers.seed_players(@tournament.players).map { |p| p.id }
+    seeded_participants_id_array.shuffle! if params[:randomly] == "true"
+    seeded_participants_id_array.each_with_index do |id, i|
+      tournament_registrations.find_by(player_id: id).set_list_position(i+1)
+    end
+    redirect_to tournament_path(id: @tournament.id, anchor: 'players')
   end
 
   private
@@ -554,11 +638,16 @@ class TournamentsController < ApplicationController
       @tournament = Tournament.find(params[:id])
     end
 
-    def check_if_admin
-      unless current_user.is_admin
+    def authenticate_admin!
+      unless current_user.present? && current_user.admin?
         respond_to do |format|
-          format.html { redirect_to @tournament, alert: 'Unauthorized! You must be an administrator to make this action.' }
-          format.json { render json: @tournament.errors, status: :unauthorized }
+          if @tournament.present?
+            format.html { redirect_to @tournament, alert: t('flash.alert.unauthorized') }
+            format.json { render json: @tournament.errors, status: :unauthorized }
+          else
+            format.html { redirect_to tournaments_path, alert: t('flash.alert.unauthorized') }
+            format.json { render json: {}, status: :unauthorized }
+          end
         end
       end
     end
@@ -568,7 +657,7 @@ class TournamentsController < ApplicationController
       date = Time.new(tp['date(1i)'], tp['date(2i)'], tp['date(3i)'],  tp['date(4i)'],  tp['date(5i)'])
       registration_deadline = Time.new(tp['registration_deadline(1i)'], tp['registration_deadline(2i)'], tp['registration_deadline(3i)'],  tp['registration_deadline(4i)'],  tp['registration_deadline(5i)'])
       if registration_deadline >= date
-        @tournament.errors.add(:registration_deadline, "must be less than the tournament start date")
+        @tournament.errors.add(:registration_deadline, t('tournaments.errors.registration_deadline'))
         return false
       else
         return true
@@ -585,12 +674,17 @@ class TournamentsController < ApplicationController
         :location, :description, :registration_fee, :total_seats,
         :host_username, :setup, :started, :finished, :active, :created_at,
         :updated_at, :subtype, :city, :end_date, :external_registration_link,
-        :total_needed_game_stations, :min_needed_registrations,
-        :is_registration_allowed, :number_of_pools)
+        :total_needed_game_stations, :min_needed_registrations, :ranking_string,
+        :is_registration_allowed, :number_of_pools, :image_link, :image_height,
+        :image_width, :canton, :federal_state, :region)
     end
 
     def set_challonge_username_and_api_key
       host_user = @tournament.host
+      # workaround for outdated ruby gem
+      Challonge::Tournament.site = "https://api.challonge.com/v1"
+      Challonge::Participant.site = "https://api.challonge.com/v1/tournaments/:tournament_id/"
+      Challonge::Match.site = "https://api.challonge.com/v1/tournaments/:tournament_id/"
       if host_user.present? and host_user.challonge_username.present? and host_user.challonge_api_key.present?
         Challonge::API.username = host_user.challonge_username
         Challonge::API.key = host_user.challonge_api_key
@@ -606,4 +700,28 @@ class TournamentsController < ApplicationController
       end
     end
 
+    def reorder_for_pools(seeded_participants_array, number_of_pools)
+      players_per_pool = (seeded_participants_array.size.to_f/number_of_pools).ceil
+      pools_hash = Hash.new
+      number_of_pools.times do |n|
+        pools_hash[n] = Array.new
+      end
+      i = 0
+      (players_per_pool.to_f/2).ceil.times do
+        number_of_pools.times do |n|
+          pools_hash[n%number_of_pools] << seeded_participants_array[i]
+          i += 1
+        end
+        number_of_pools.times do |n|
+          pools_hash[(number_of_pools-1-n)%number_of_pools] << seeded_participants_array[i]
+          i += 1
+        end
+      end
+      seeded_participants_array = []
+      number_of_pools.times do |n|
+        pools_hash[n].pop if pools_hash[n].size > players_per_pool
+        seeded_participants_array += pools_hash[n]
+      end
+      return seeded_participants_array
+    end
 end
