@@ -2,8 +2,9 @@ class UserMailer < ApplicationMailer
 
   def welcome_email
     @user = params[:user]
-    @url  = 'https://ssb-tournament-manager.herokuapp.com/tournaments'
-    mail(to: @user.email, subject: 'Welcome to the SwissSmash Tournament Manager')
+    @locale = get_locale(@user.country_code)
+    @url = tournaments_url(@user.country_code, '')
+    mail(to: @user.email, from: from(@user.country_code), subject: t("users.mailer.welcome_#{@user.country_code}", locale: @locale))
   end
 
 end
