@@ -4,16 +4,32 @@ class Tournament < ApplicationRecord
   has_many :matches, dependent: :destroy
   has_many :results, dependent: :destroy
 
-  validates :name, uniqueness: true
+  validates :name, uniqueness: true, presence: true
 
-  scope :active_2019, -> { where(active: true).where('date > ? AND date < ?', Time.local(2019,1,1), Time.local(2019,12,31,23,59,59)) }
+  scope :all_from, ->(country_code) { where(country_code: country_code) }
+  scope :active, -> { where(active: true) }
   scope :upcoming, -> { where('date > ?', Time.now) }
-  scope :ongoing, -> { where('date <= ? AND date >= ?', Time.now, Time.now - 6.hours) }
-  scope :past, -> { where('date < ?', Time.now - 6.hours) }
+  scope :upcoming_with_today, -> { where('date >= ?', Time.now.beginning_of_day) }
+  scope :ongoing, -> { where('(finished IS FALSE OR finished IS NULL) AND date <= ? AND date >= ?', Time.now, Time.now - 6.hours) }
+  scope :past, -> { where('started AND finished OR date < ?', Time.now - 6.hours) }
   scope :for_calendar, -> { where(active: true).where('date > ? AND date < ?', 2.weeks.ago, Date.today + 4.months) }
   scope :from_city, -> (city) { where("name ILIKE ? OR name ILIKE ? OR location ILIKE ? OR location ILIKE ?", "%#{ActiveRecord::Base.sanitize_sql_like(city)}%", "%#{ActiveRecord::Base.sanitize_sql_like(city.downcase)}%", "%#{ActiveRecord::Base.sanitize_sql_like(city)}%", "%#{ActiveRecord::Base.sanitize_sql_like(city.downcase)}%") }
 
+  before_create :set_canton
+  before_create :set_federal_state
+  before_create :set_region
+
   MAX_PAST_TOURNAMENTS_PER_PAGE = 20
+
+  def self.search(search)
+    if search
+      sanitizedSearch = ActiveRecord::Base.sanitize_sql_like(search)
+      # name location city ranking_string
+      where("name ILIKE ? or location ILIKE ? or city ILIKE ?", "%#{sanitizedSearch}%", "%#{sanitizedSearch}%", "%#{sanitizedSearch}%")
+    else
+      :all
+    end
+  end
 
   def cancelled?
     !self.started and self.finished
@@ -24,7 +40,11 @@ class Tournament < ApplicationRecord
   end
 
   def has_pools?
-    self.number_of_pools > 0
+    self.number_of_pools.to_i > 0
+  end
+
+  def weekly?
+    self.subtype == 'weekly'
   end
 
   def game_stations_count
@@ -37,6 +57,175 @@ class Tournament < ApplicationRecord
 
   def host
     User.find_by(username: self.host_username) if self.host_username.present?
+  end
+
+  def set_canton
+    return unless self.country_code == 'ch'
+    return if self.canton.present?
+    cantons_raw = ApplicationController.helpers.cantons_raw
+    cantons_de = I18n.t(cantons_raw, scope: 'defines.cantons', locale: :de).map(&:downcase)
+    cantons_fr = I18n.t(cantons_raw, scope: 'defines.cantons', locale: :fr).map(&:downcase)
+    cantons_en = I18n.t(cantons_raw, scope: 'defines.cantons', locale: :en).map(&:downcase)
+    # First: Try to determine canton from city
+    if self.city.present?
+      city = self.city.downcase
+      city = city.gsub('basel', 'basel-stadt').gsub('bâle', 'bâle-ville').gsub('gallen', 'st. gallen')
+      if (cantons_de.include?(city) || cantons_fr.include?(city) || cantons_en.include?(city))
+        self.canton = cantons_raw[cantons_de.index(city)] if cantons_de.index(city).present?
+        self.canton = cantons_raw[cantons_fr.index(city)] if cantons_fr.index(city).present?
+        self.canton = cantons_raw[cantons_en.index(city)] if cantons_en.index(city).present?
+        return # as soon as canton was found
+      end
+    end
+    # Second: Try to determine canton from a word in location
+    if self.location.present?
+      self.location.downcase.split(' ').each do |l|
+        l = l.gsub(',', '').gsub('basel', 'basel-stadt').gsub('bâle', 'bâle-ville').gsub('gallen', 'st. gallen')
+        if (cantons_de.include?(l) || cantons_fr.include?(l) || cantons_en.include?(l))
+          self.canton = cantons_raw[cantons_de.index(l)] if cantons_de.index(l).present?
+          self.canton = cantons_raw[cantons_fr.index(l)] if cantons_fr.index(l).present?
+          self.canton = cantons_raw[cantons_en.index(l)] if cantons_en.index(l).present?
+          return # as soon as canton was found
+        end
+      end
+      # Third: Try to find the canton with the help of Google Maps
+      require 'open-uri'
+      require 'json'
+      begin
+        json_data = JSON.parse(URI.open("https://maps.googleapis.com/maps/api/geocode/json?address=#{ERB::Util.url_encode(self.location)}&components=country:CH&key=#{ENV['GOOGLE_MAPS_SERVER_SIDE_API_KEY']}&outputFormat=json").read)
+        if json_data["status"] == "OK" && json_data["results"].present? && json_data["results"][0].present?
+          json_data["results"][0]["address_components"].each do |res|
+            if (res["types"].present? && res["types"].include?('administrative_area_level_1'))
+              if res["long_name"].present?
+                ln = res["long_name"].downcase
+                if (cantons_de.include?(ln) || cantons_fr.include?(ln) || cantons_en.include?(ln))
+                  self.canton = cantons_raw[cantons_de.index(ln)] if cantons_de.index(ln).present?
+                  self.canton = cantons_raw[cantons_fr.index(ln)] if cantons_fr.index(ln).present?
+                  self.canton = cantons_raw[cantons_en.index(ln)] if cantons_en.index(ln).present?
+                  return # as soon as canton was found
+                end
+              end
+            end
+          end
+        end
+      rescue OpenURI::HTTPError => ex
+        puts ex
+      end
+    end
+  end
+
+  def set_federal_state
+    return unless self.country_code == 'de'
+    return if self.federal_state.present?
+    federal_states_raw = ApplicationController.helpers.federal_states_raw
+    federal_states_de = I18n.t(federal_states_raw, scope: 'defines.federal_states', locale: :de).map(&:downcase)
+    federal_states_en = I18n.t(federal_states_raw, scope: 'defines.federal_states', locale: :en).map(&:downcase)
+    # First: Try to determine federal_state from city
+    if self.city.present?
+      city = self.city.downcase
+      #city = city.gsub('basel', 'basel-stadt').gsub('bâle', 'bâle-ville').gsub('gallen', 'st. gallen')
+      if (federal_states_de.include?(city) || federal_states_en.include?(city))
+        self.federal_state = federal_states_raw[federal_states_de.index(city)] if federal_states_de.index(city).present?
+        self.federal_state = federal_states_raw[federal_states_en.index(city)] if federal_states_en.index(city).present?
+        return # as soon as federal_state was found
+      end
+    end
+    # Second: Try to determine federal_state from a word in location
+    if self.location.present?
+      self.location.downcase.split(' ').each do |l|
+        l = l.gsub(',', '')#.gsub('basel', 'basel-stadt').gsub('bâle', 'bâle-ville').gsub('gallen', 'st. gallen')
+        if (federal_states_de.include?(l) || federal_states_en.include?(l))
+          self.federal_state = federal_states_raw[federal_states_de.index(l)] if federal_states_de.index(l).present?
+          self.federal_state = federal_states_raw[federal_states_en.index(l)] if federal_states_en.index(l).present?
+          return # as soon as federal_state was found
+        end
+      end
+      # Third: Try to find the federal_state with the help of Google Maps
+      require 'open-uri'
+      require 'json'
+      begin
+        json_data = JSON.parse(URI.open("https://maps.googleapis.com/maps/api/geocode/json?address=#{ERB::Util.url_encode(self.location)}&components=country:DE&key=#{ENV['GOOGLE_MAPS_SERVER_SIDE_API_KEY']}&outputFormat=json").read)
+        if json_data["status"] == "OK" && json_data["results"].present? && json_data["results"][0].present?
+          json_data["results"][0]["address_components"].each do |res|
+            if (res["types"].present? && res["types"].include?('administrative_area_level_1'))
+              if res["long_name"].present?
+                sn = res["short_name"]
+                if federal_states_raw.include?(sn)
+                  self.federal_state = sn
+                  return # as soon as federal_state was found
+                end
+                # long_name will most likely never be necessary
+                ln = res["long_name"].downcase
+                if (federal_states_de.include?(ln) || federal_states_en.include?(ln))
+                  self.federal_state = federal_states_raw[federal_states_de.index(ln)] if federal_states_de.index(ln).present?
+                  self.federal_state = federal_states_raw[federal_states_en.index(ln)] if federal_states_en.index(ln).present?
+                  return # as soon as federal_state was found
+                end
+              end
+            end
+          end
+        end
+      rescue OpenURI::HTTPError => ex
+        puts ex
+      end
+    end
+  end
+
+  def set_region
+    return unless self.country_code == 'fr'
+    return if self.region.present?
+    regions_raw = ApplicationController.helpers.regions_raw
+    regions_fr = I18n.t(regions_raw, scope: 'defines.regions', locale: :fr).map(&:downcase)
+    regions_en = I18n.t(regions_raw, scope: 'defines.regions', locale: :en).map(&:downcase)
+    # First: Try to determine region from city
+    if self.city.present?
+      city = self.city.downcase
+      #city = city.gsub('basel', 'basel-stadt').gsub('bâle', 'bâle-ville').gsub('gallen', 'st. gallen')
+      if (regions_fr.include?(city) || regions_en.include?(city))
+        self.region = regions_raw[regions_fr.index(city)] if regions_fr.index(city).present?
+        self.region = regions_raw[regions_en.index(city)] if regions_en.index(city).present?
+        return # as soon as region was found
+      end
+    end
+    # Second: Try to determine region from a word in location
+    if self.location.present?
+      self.location.downcase.split(' ').each do |l|
+        l = l.gsub(',', '')#.gsub('basel', 'basel-stadt').gsub('bâle', 'bâle-ville').gsub('gallen', 'st. gallen')
+        if (regions_fr.include?(l) || regions_en.include?(l))
+          self.region = regions_raw[regions_fr.index(l)] if regions_fr.index(l).present?
+          self.region = regions_raw[regions_en.index(l)] if regions_en.index(l).present?
+          return # as soon as region was found
+        end
+      end
+      # Third: Try to find the region with the help of Google Maps
+      require 'open-uri'
+      require 'json'
+      begin
+        json_data = JSON.parse(URI.open("https://maps.googleapis.com/maps/api/geocode/json?address=#{ERB::Util.url_encode(self.location)}&components=country:FR&key=#{ENV['GOOGLE_MAPS_SERVER_SIDE_API_KEY']}&outputFormat=json").read)
+        if json_data["status"] == "OK" && json_data["results"].present? && json_data["results"][0].present?
+          json_data["results"][0]["address_components"].each do |res|
+            if (res["types"].present? && res["types"].include?('administrative_area_level_1'))
+              if res["long_name"].present?
+                sn = res["short_name"]
+                if regions_raw.include?(sn)
+                  self.region = sn
+                  return # as soon as region was found
+                end
+                # long_name will most likely never be necessary
+                ln = res["long_name"].downcase
+                if (regions_fr.include?(ln) || regions_en.include?(ln))
+                  self.region = regions_raw[regions_fr.index(ln)] if regions_fr.index(ln).present?
+                  self.region = regions_raw[regions_en.index(ln)] if regions_en.index(ln).present?
+                  return # as soon as region was found
+                end
+              end
+            end
+          end
+        end
+      rescue OpenURI::HTTPError => ex
+        puts ex
+      end
+    end
   end
 
 end
